backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.38
psycopg[binary]==3.2.4
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
PyJWT==2.10.1

